# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meridian API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Start Run",
        "operationId": "start_run_api_v1_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartRunResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/events": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run Events",
        "operationId": "get_run_events_api_v1_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "last_event_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last Event Id"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/EventEnvelope"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CitationView": {
        "properties": {
          "corpus": {
            "type": "string",
            "title": "Corpus"
          },
          "section": {
            "type": "string",
            "title": "Section"
          },
          "claim": {
            "type": "string",
            "title": "Claim"
          },
          "applied_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied To"
          },
          "quoted": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quoted"
          },
          "exists": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exists"
          },
          "supports_claim": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supports Claim"
          },
          "entailed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entailed"
          },
          "verified": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified"
          },
          "qc_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qc Note"
          }
        },
        "type": "object",
        "required": [
          "corpus",
          "section",
          "claim"
        ],
        "title": "CitationView"
      },
      "CompensationView": {
        "properties": {
          "step": {
            "type": "string",
            "title": "Step"
          },
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "detail": {
            "additionalProperties": true,
            "type": "object",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "step",
          "outcome",
          "note"
        ],
        "title": "CompensationView"
      },
      "ConditionView": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "citation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citation"
          },
          "raised_by": {
            "type": "string",
            "title": "Raised By"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "description",
          "raised_by"
        ],
        "title": "ConditionView"
      },
...
      "Decision": {
        "type": "string",
        "enum": [
          "approved",
          "approved_with_conditions",
          "suspended",
          "denied",
          "incomplete"
        ],
        "title": "Decision"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorEnvelope"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "active_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Run Id"
          },
          "active_evaluation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Evaluation Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "EventEnvelope": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "loan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loan Id"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "line": {
            "type": "string",
            "title": "Line"
          }
        },
        "type": "object",
        "required": [
          "id",
          "run_id",
          "type",
          "actor",
          "timestamp",
          "line"
        ],
        "title": "EventEnvelope"
      },
      "FindingView": {
        "properties": {
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "severity",
          "kind",
          "detail"
        ],
        "title": "FindingView"
      },
      "GuidelineFindingView": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "rounds_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rounds Used"
          },
          "sufficient": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sufficient"
          },
          "citations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Citations"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "GuidelineFindingView"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IdempotencyView": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "calls_before_second_run": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calls Before Second Run"
          },
          "calls_after_second_run": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calls After Second Run"
          }
        },
        "type": "object",
        "title": "IdempotencyView"
      },
...
      "LedgerEntryView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "step": {
            "type": "string",
            "title": "Step"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "detail": {
            "additionalProperties": true,
            "type": "object",
            "title": "Detail"
          },
          "timestamp": {
            "type": "number",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "step",
          "phase",
          "outcome",
          "timestamp"
        ],
        "title": "LedgerEntryView"
      },
      "LoanRunResult": {
        "properties": {
          "loan_id": {
            "type": "string",
            "title": "Loan Id"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "scenario_name": {
            "type": "string",
            "title": "Scenario Name"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Decision"
              },
              {
                "type": "null"
              }
            ]
          },
          "decision_rationale": {
            "type": "string",
            "title": "Decision Rationale",
            "default": ""
          },
          "expected_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Decision"
          },
          "expected_matches": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Matches"
          },
          "overlays": {
            "items": {
              "$ref": "#/components/schemas/OverlayConflictView"
            },
            "type": "array",
            "title": "Overlays"
          },
          "value_acceptance": {
            "$ref": "#/components/schemas/ValueAcceptanceView"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/ConditionView"
            },
            "type": "array",
            "title": "Conditions"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationView"
            },
            "type": "array",
            "title": "Citations"
          },
          "qc_findings": {
            "items": {
              "$ref": "#/components/schemas/FindingView"
            },
            "type": "array",
            "title": "Qc Findings"
          },
          "controls_fired": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Controls Fired"
          },
          "guideline_findings": {
            "items": {
              "$ref": "#/components/schemas/GuidelineFindingView"
            },
            "type": "array",
            "title": "Guideline Findings"
          },
          "ledger": {
            "items": {
              "$ref": "#/components/schemas/LedgerEntryView"
            },
            "type": "array",
            "title": "Ledger"
          },
          "compensation": {
            "items": {
              "$ref": "#/components/schemas/CompensationView"
            },
            "type": "array",
            "title": "Compensation"
          },
          "notices": {
            "items": {
              "$ref": "#/components/schemas/NoticeView"
            },
            "type": "array",
            "title": "Notices"
          },
          "cycle_time": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CycleTimeView"
              },
              {
                "type": "null"
              }
            ]
          },
          "agent_runs": {
            "items": {
              "$ref": "#/components/schemas/AgentRunView"
            },
            "type": "array",
            "title": "Agent Runs"
          },
          "tri_merge_calls": {
            "type": "integer",
            "title": "Tri Merge Calls",
            "default": 0
          },
          "idempotency": {
            "$ref": "#/components/schemas/IdempotencyView"
          }
        },
        "type": "object",
        "required": [
          "loan_id",
          "scenario_name"
        ],
        "title": "LoanRunResult"
      },
      "NoticeView": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "issued_on": {
            "type": "string",
            "title": "Issued On"
          },
          "due_on": {
            "type": "string",
            "title": "Due On"
          },
          "citation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citation"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "issued_on",
          "due_on"
        ],
        "title": "NoticeView"
      },
      "OverlayConflictView": {
        "properties": {
          "dimension": {
            "type": "string",
            "title": "Dimension"
          },
          "actual": {
            "title": "Actual"
          },
          "agency": {
            "title": "Agency"
          },
          "overlay": {
            "title": "Overlay"
          },
          "citation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citation"
          },
          "exception_granted": {
            "type": "boolean",
            "title": "Exception Granted",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "dimension"
        ],
        "title": "OverlayConflictView"
      },
      "RunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "completed",
              "awaiting_approval",
              "failed"
            ],
            "title": "Status"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "scenario": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScenarioOption"
              },
              {
                "type": "null"
              }
            ]
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventEnvelope"
            },
            "type": "array",
            "title": "Events"
          },
          "trace_lines": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Trace Lines"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LoanRunResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "approval": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval"
          },
          "error_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Type"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "terminal_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal At"
          },
          "terminal_persistence_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal Persistence Error"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status"
        ],
        "title": "RunResponse",
        "description": "A reconnectable snapshot consumed by the case workspace."
      },
      "ScenarioOption": {
        "properties": {
          "scenario_id": {
            "type": "integer",
            "title": "Scenario Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "loan_id": {
            "type": "string",
            "title": "Loan Id"
          },
          "expected_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Decision"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "scenario_id",
          "label",
          "name",
          "loan_id"
        ],
        "title": "ScenarioOption"
      },
...
      "StartRunRequest": {
        "properties": {
          "scenario_id": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Scenario Id"
          },
          "judgment_kind": {
            "type": "string",
            "enum": [
              "stub",
              "crew"
            ],
            "title": "Judgment Kind",
            "default": "stub"
          },
          "roster": {
            "type": "string",
            "enum": [
              "production",
              "demo"
            ],
            "title": "Roster",
            "default": "production"
          },
          "policy_attack": {
            "type": "boolean",
            "title": "Policy Attack",
            "default": false
          },
          "submit_twice": {
            "type": "boolean",
            "title": "Submit Twice",
            "default": false
          },
          "auto_approve": {
            "type": "boolean",
            "title": "Auto Approve",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scenario_id"
        ],
        "title": "StartRunRequest"
      },
      "StartRunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "const": "running",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status"
        ],
        "title": "StartRunResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValueAcceptanceView": {
        "properties": {
          "offered": {
            "type": "boolean",
            "title": "Offered",
            "default": false
          },
          "exercised": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exercised"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale",
            "default": ""
          }
        },
        "type": "object",
        "title": "ValueAcceptanceView"
      }
    }
  }
}